requirements.txt
torch>=2.0.0
langchain-openai>=0.0.5
langgraph>=0.0.10
python-dotenv>=1.0.0
typing-extensions>=4.5.0 